Replace usage of JToken.ToString()
In Newtonsoft.Json v3.5, JToken.ToString() returned the raw JSON
representation of the token. In later versions of
Newtonsoft.Json, JToken.ToString() returns a simple string
representation of the value. See the examples below:
- v3.5: "\"Hello World\""
- Later versions: "Hello World"

In this commit, I've updated the project to work with later versions
of Newtonsoft.Json as well as v3.5. I've replaced some usages of
JToken.ToString(). When we need the raw JSON representation, we use
JsonConvert.Serialize(). When we need the string value of a string
JToken, we use JToken.Value<string>().