lang/csharp: Throw ArgumentNullException in SpecificDatumWriter
Before this change, SpecificDatumWriter would throw an AvroTypeException stating that the provied value
was the wrong type when the value was actually null. This was problematic because it gave the developer
an impression that the cause of the error was different than it actually was. To clarify the error,
the method throws an ArgumentNullException instead in this scenario.